+++ .eslintrc.js
@@ -6,12 +6,14 @@ module.exports = {
   overrides: [
     {
       files: ['*.ts', '*.tsx'],
+
       rules: {
         '@typescript-eslint/no-shadow': ['off'],
         'no-shadow': 'off',
         'no-undef': 'off',
         '@typescript-eslint/no-unused-vars': ['warn'],
         'react-native/no-inline-styles': 'off',
+        'react-hooks/exhaustive-deps': 'warn',
       },
     },
   ],